add method extract,kickOut for SkM and SkAnyM. add some useful matrix methods for SkMat

+++ bgfx_study/app/src/main/cpp/thirds/lua/IMemory.cpp
@@ -724,8 +724,8 @@ void MemoryUtils::write(void *data, const char t, size_t bytesIndex, double val)
     }
 }
 
-const int MemoryUtils::computeBytesIndex(const char *types, int index) {
-    int bytesIndex = 0;
+const size_t MemoryUtils::computeBytesIndex(const char *types, size_t index) {
+    size_t bytesIndex = 0;
     const size_t len = strlen(types);
     for (int i = 0; i < index; ++i) {
         bytesIndex += MemoryUtils::getUnitSize(types[i % len]);
